requirements.txt
httpx>=0.27.0
websockets>=12.0
numpy>=1.26.0
scipy>=1.12.0
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
rich>=13.0.0
